Add Options and Fallback arguments to Properties
Add optional Options and Fallback arguments that lets you customize each
property or use default options for multiple properties.

Instead of defining options for every parameter, it's often convenient
to define it once. Alternative implementation is a fallback key in
Options but that risks that the user picks that key.

This gives you access to slider widgets, limits, etc:

    Slab.Properties(trails, {
        -- integers
        history       = {MinNumber = trails.display, MaxNumber = 1000},
        display       = {MinNumber = 0,              MaxNumber = trails.history},
        -- floats
        initial_scale = {MinNumber = 0.001,          MaxNumber = 5},
        scale_down    = {MinNumber = 0.001,          MaxNumber = 1},
        alpha         = {MinNumber = 0.001,          MaxNumber = 1},
    },
                        {MinNumber = -200.1,         MaxNumber = 200})